requirements.txt
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==7.4.0
reportlab==4.0.7
Pillow==10.1.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0